pyproject.toml
[project]
name = "facade-grammar"
version = "0.1.0"
description = "Procedural grammar extraction from Dutch canal houses (3D BAG + OSM + Mapillary + SAM 3)."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sf-hamilton[rich,visualization]>=1.89,<2",
    "pydantic>=2.13,<3",
    "pydantic-settings>=2.13,<3",
    "pyyaml>=6.0,<7",
    "httpx>=0.28,<0.29",
    "osmnx>=2.1,<3",
    "polars>=1.39,<2",
    "shapely>=2.1,<3",
    "matplotlib>=3.10,<4",
    "pyproj>=3.7,<4",
    "typer>=0.24,<0.25",
    "numpy>=2.2,<3",
    "pillow>=11.0,<12",
    "pandera[polars]>=0.31,<0.32",
    "tenacity>=9.1,<10",
]

[project.scripts]
facade-grammar = "facade_grammar.cli:cli"

[dependency-groups]
dev = [
    "ruff>=0.15,<0.16",
    "ty==0.0.31",
    "pytest>=9.0,<10",
    "pytest-randomly>=4.1,<5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/facade_grammar"]

[tool.uv.workspace]
members = ["services/*"]

[tool.ruff]
src = ["src", "scripts", "tests"]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # bugbear
    "A",   # builtins-shadowing
    "C4",  # comprehensions
    "SIM", # simplify
    "RUF", # ruff-specific
]

[tool.ty.environment]
root = ["./src"]
python-version = "3.14"
python = "./.venv"

[tool.ty.src]
exclude = ["services/**"]  # sam-service has its own venv / type-check pass

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra --strict-markers --strict-config --import-mode=importlib"
filterwarnings = ["error"]
